footer.html:add condition for time display
We don't want the time-stamped on every page. That would mean pushing
every page every time. That happens right now because the design is
changing so frequently, but after that stabilizes, not every page will
get changed.

Why doesn't this work? I couldn't even get it to print {{ page.title }}