footer.html:added RSS feeds (see comment) (closes #93)
I like it, but it could use a bit more design. Maybe @bloopletech
would accept PRs so templates could be added. The rss2html project
is at https://github.com/bloopletech/rss2html

Generally speaking, these feeds might be moved onto a separate page, or
something else may be done with it. We'll try this for a bit and see how
we like it.