update jekyll-paginate-v2 to 1.94
I was notified by @sverrirs through a ticket
(sverrirs/jekyll-paginate-v2#20) about an update.

I had left a bug report on that ticket on Oct. 30, but I don't recall
for certain if the problem was related to this plug-in. We've been using
it now ever since we set up the site and it's been fitting our needs
well.

Though we haven't solved this issue yet:

**pagination works on category pages but not on tag pages**

Though, tbh, we haven't been actively working on it.

Anyway, I updated the plug-in and rebuilt the site. No changed pages to
push, so just pushing the Gemfile that has the version change.

@erichoog, no reason to update Gemfile, only the .lock file.. as far as
you know?